refactor: separate reducers
